refactor: pytest-docker fixtures
